site/wwwroot/requirements.txt
firebase-admin==6.5.0
pandas==2.2.2
numpy==1.26.4
requests==2.32.3
sqlalchemy==2.0.31
psycopg2-binary==2.9.5
azure-functions==1.20.0